# src/main/resources/static/swagger.yml
openapi: 3.0.4
info:
  title: itPlace API Document
  description: |-
    지도 기반 내 주변 LG유플러스 멤버십 혜택 정보 조회 서비스 'itPlace'의 API 명세서입니다.
    
    모든 API 응답은 표준 응답 포맷(ApiResponse)을 따릅니다.
  version: 1.0.0
servers:
  - url: https://userapi.itplace.click/api/v1
    description: 개발 서버

  - url: http://localhost:8080/api/v1
    description: 로컬 테스트 서버
...
paths:
...
  /maps/nearby:
    get:
      tags:
        - Map
      summary: 사용자 위치 기반 전체 지점 목록 조회
      description: 사용자의 위치(위도, 경도)를 기반으로 지정된 반경 내의 모든 지점을 조회합니다.
      parameters:
        - name: lat
          in: query
          required: true
          description: 위도 좌표
          schema:
            type: number
            format: double
            example: 37.5665
        - name: lng
          in: query
          required: true
          description: 경도 좌표
          schema:
            type: number
            format: double
            example: 126.9780
        - name: radiusMeters
          in: query
          required: true
          description: 검색 반경 (미터 단위)
          schema:
            type: number
            format: double
            minimum: 0
            example: 1000
      responses:
        '200':
          description: 전체 지점 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StoreDetailDto'
              examples:
                success:
                  summary: 전체 지점 목록 조회 성공 예시
                  value:
                    code: STORE_LIST_SUCCESS
                    status: OK
                    message: 지점 목록 조회에 성공했습니다.
                    data:
                      - store:
                          storeId: 1
                          storeName: "스타벅스 강남점"
                          business: "커피전문점"
                          city: "서울특별시"
                          town: "강남구"
                          legalDong: "역삼동"
                          address: "서울특별시 강남구 역삼동 123-45"
                          roadName: "강남대로"
                          roadAddress: "서울특별시 강남구 강남대로 456"
                          postCode: "06234"
                        partner:
                          partnerId: 101
                          partnerName: "스타벅스"
                          image: "https://example.com/starbucks-logo.png"
                          category: "카페"
                        tierBenefit:
                          - grade: "BASIC"
                            context: "5% 할인"
                          - grade: "VIP"
                            context: "10% 할인"
                          - grade: "VVIP"
                            context: "15% 할인"
                        distance: 0.5
                    timestamp: '2025-07-15T11:00:00.000Z'
  /maps/nearby/category:
    get:
      tags:
        - Map
      summary: 사용자 위치 기반 카테고리별 지점 목록 조회
      description: 사용자의 위치를 기반으로 지정된 반경 내의 특정 카테고리 지점을 조회합니다. 카테고리가 null, 빈 문자열, 또는 "전체"인 경우 모든 지점을 반환합니다.
      parameters:
        - name: lat
          in: query
          required: true
          description: 위도 좌표
          schema:
            type: number
            format: double
            example: 37.5665
        - name: lng
          in: query
          required: true
          description: 경도 좌표
          schema:
            type: number
            format: double
            example: 126.9780
        - name: radiusMeters
          in: query
          required: true
          description: 검색 반경 (미터 단위)
          schema:
            type: number
            format: double
            minimum: 0
            example: 1000
        - name: category
          in: query
          required: false
          description: 카테고리 필터 (null, 빈 문자열, 또는 "전체"이면 모든 카테고리 반환)
          schema:
            type: string
            example: "카페"
      responses:
        '200':
          description: 카테고리별 지점 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StoreDetailDto'
              examples:
                success:
                  summary: 카페 카테고리 지점 목록 조회 성공 예시
                  value:
                    code: STORE_LIST_SUCCESS
                    status: OK
                    message: 지점 목록 조회에 성공했습니다.
                    data:
                      - store:
                          storeId: 1
                          storeName: "스타벅스 강남점"
                          business: "커피전문점"
                          city: "서울특별시"
                          town: "강남구"
                          legalDong: "역삼동"
                          address: "서울특별시 강남구 역삼동 123-45"
                          roadName: "강남대로"
                          roadAddress: "서울특별시 강남구 강남대로 456"
                          postCode: "06234"
                        partner:
                          partnerId: 101
                          partnerName: "스타벅스"
                          image: "https://example.com/starbucks-logo.png"
                          category: "카페"
                        tierBenefit:
                          - grade: "BASIC"
                            context: "5% 할인"
                          - grade: "VIP"
                            context: "10% 할인"
                          - grade: "VVIP"
                            context: "15% 할인"
                        distance: 0.5
                      - store:
                          storeId: 2
                          storeName: "투썸플레이스 역삼점"
                          business: "커피전문점"
                          city: "서울특별시"
                          town: "강남구"
                          legalDong: "역삼동"
                          address: "서울특별시 강남구 역삼동 678-90"
                          roadName: "테헤란로"
                          roadAddress: "서울특별시 강남구 테헤란로 123"
                          postCode: "06233"
                        partner:
                          partnerId: 102
                          partnerName: "투썸플레이스"
                          image: "https://example.com/twosome-logo.png"
                          category: "카페"
                        tierBenefit:
                          - grade: "BASIC"
                            context: "3% 할인"
                          - grade: "VIP"
                            context: "5% 할인"
                          - grade: "VVIP"
                            context: "5% 할인"
                        distance: 0.8
                    timestamp: '2025-07-15T11:00:00.000Z'
  /maps/nearby/search:
    get:
      tags:
        - Map
      summary: 사용자 위치 기반 키워드로 지점 검색
      description: 사용자의 위치를 기반으로 지정된 반경 내의 지점 중 키워드를 기반으로 검색합니다. 키워드가 지점이름, 제휴사이름, 직종, 카테고리 중 하나라도 포함되어있으면 조회됩니다.
      parameters:
        - name: lat
          in: query
          required: true
          description: 위도 좌표
          schema:
            type: number
            format: double
            example: 37.5665
        - name: lng
          in: query
          required: true
          description: 경도 좌표
          schema:
            type: number
            format: double
            example: 126.9780
        - name: category
          in: query
          required: false
          description: 카테고리 필터 (선택사항)
          schema:
            type: string
            example: "카페"
        - name: keyword
          in: query
          required: true
          description: 검색 키워드 (지점명, 제휴사명, 업종, 카테고리에서 검색)
          schema:
            type: string
            example: "스타벅스"
      responses:
        '200':
          description: 키워드 검색 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StoreDetailDto'
              examples:
                success:
                  summary: 스타벅스 키워드 검색 성공 예시
                  value:
                    code: STORE_LIST_SUCCESS
                    status: OK
                    message: 지점 목록 조회에 성공했습니다.
                    data:
                      - store:
                          storeId: 1
                          storeName: "스타벅스 강남점"
                          business: "커피전문점"
                          city: "서울특별시"
                          town: "강남구"
                          legalDong: "역삼동"
                          address: "서울특별시 강남구 역삼동 123-45"
                          roadName: "강남대로"
                          roadAddress: "서울특별시 강남구 강남대로 456"
                          postCode: "06234"
                        partner:
                          partnerId: 101
                          partnerName: "스타벅스"
                          image: "https://example.com/starbucks-logo.png"
                          category: "카페"
                        tierBenefit:
                          - grade: "BASIC"
                            context: "5% 할인"
                          - grade: "VIP"
                            context: "10% 할인"
                          - grade: "VVIP"
                            context: "15% 할인"
                        distance: 0.5
                    timestamp: '2025-07-15T11:00:00.000Z'
        '400':
          description: 키워드 누락 또는 잘못된 요청 파라미터
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                keyword_required:
                  summary: 키워드 누락 에러
                  value:
                    code: KEYWORD_REQUIRED
                    status: BAD_REQUEST
                    message: 키워드를 입력해주세요
                    data: null
                    timestamp: '2025-07-15T11:00:00.000Z'
  /maps/nearby/itplace-ai:
    get:
      tags:
        - Map
      summary: 잇플 ai가 추천한 제휴처와 사용자 위치 기반으로 매장을 조회
      description: 제휴처 이름으로 제휴처 id를 얻어 매장을 검색합니다.
      parameters:
        - name: lat
          in: query
          required: true
          description: 위도 좌표
          schema:
            type: number
            format: double
            example: 37.5665
        - name: lng
          in: query
          required: true
          description: 경도 좌표
          schema:
            type: number
            format: double
            example: 126.9780
        - name: partnerName
          in: query
          required: true
          description: 제휴사 이름
          schema:
            type: string
            example: "롯데월드"
      responses:
        '200':
          description: 매장 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StoreDetailDto'
              examples:
                success:
                  summary: 롯데월드 조회 성공 예시
                  value:
                    code: STORE_LIST_SUCCESS
                    status: OK
                    message: 사용자 위치 기반 매장 조회에 성공했습니다.
                    data:
                      - store:
                          storeId: 10266
                          storeName: "롯데월드 어드벤처"
                          business: "테마"
                          city: "서울특별시"
                          town: "송파구"
                          legalDong: "잠실동"
                          address: "잠실동 40-1"
                          roadName: "서울특별시 송파구 올림픽로"
                          roadAddress: "서울특별시 송파구 올림픽로 240"
                          postCode: "05554"
                          longitude: 127.097184
                          latitude: 37.511244
                        partner:
                          partnerId: 22
                          partnerName: "롯데월드"
                          image: "https://itplacepartners.s3.ap-northeast-2.amazonaws.com/img/lotteworldadventure.jpg"
                          category: "엑티비티"
                        tierBenefit:
                          - grade: "VIP콕"
                            context: "본인 종합이용권(1 Day / AFTER 4) 42% 할인\r\n (동반 1인 30% 할인)"
                        distance: 20.5
                    timestamp: '2025-07-15T11:00:00.000Z'
                online-benefit:
                  summary: 매장이 없는 경우(온라인 혜택 같은)
                  value:
                    code: STORE_LIST_SUCCESS
                    status: OK
                    message: 사용자 위치 기반 매장 조회에 성공했습니다.
                    data: [ ]
                    timestamp: '2025-07-15T11:00:00.000Z'
        '404':
          description: 없는 제휴처를 요청할 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                keyword_required:
                  summary: 제휴처를 찾을 수 없음
                  value:
                    code: PARTNER_NOT_FOUND
                    status: NOT_FOUND
                    message: 제휴사를 찾을 수 없습니다.
                    data: null
                    timestamp: '2025-07-15T11:00:00.000Z'

  # =====================================================
  # 혜택 관련 기능 API
  # =====================================================
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: "서버에서 정의한 응답 코드"
          example: "SUCCESS_200_1"
        status:
          type: string
          description: "HTTP 상태 메시지"
          example: "OK"
        message:
          type: string
          description: "응답에 대한 설명 메시지"
          example: "요청이 정상적으로 처리되었습니다."
        data:
          type: object
          description: "실제 응답 데이터 (nullable)"
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: "응답 시간"

    # ----------------------------------------------------
    # Request DTOs
    # ----------------------------------------------------
...
    StoreDetailDto:
      type: object
      properties:
        store:
          $ref: '#/components/schemas/StoreDto'
        partner:
          $ref: '#/components/schemas/PartnerDto'
        tierBenefit:
          type: array
          items:
            $ref: '#/components/schemas/TierBenefitDto'
        distance:
          type: number
          format: double
          description: 사용자 위치로부터의 거리 (km)
          example: 0.5
    StoreDto:
      type: object
      properties:
        storeId:
          type: integer
          format: int64
          description: 지점 ID
          example: 1
        storeName:
          type: string
          description: 지점명
          example: "스타벅스 강남점"
        business:
          type: string
          description: 업종
          example: "커피전문점"
        city:
          type: string
          description: 시
          example: "서울특별시"
        town:
          type: string
          description: 구
          example: "강남구"
        legalDong:
          type: string
          description: 법정동
          example: "역삼동"
        address:
          type: string
          description: 지번주소
          example: "서울특별시 강남구 역삼동 123-45"
        roadName:
          type: string
          description: 도로명
          example: "강남대로"
        roadAddress:
          type: string
          description: 도로명주소
          example: "서울특별시 강남구 강남대로 456"
        postCode:
          type: string
          description: 우편번호
          example: "06234"
    PartnerDto:
      type: object
      properties:
        partnerId:
          type: integer
          format: int64
          description: 파트너 ID
          example: 101
        partnerName:
          type: string
          description: 파트너명
          example: "스타벅스"
        image:
          type: string
          description: 파트너 이미지 URL
          example: "https://example.com/starbucks-logo.png"
        category:
          type: string
          description: 카테고리
          example: "카페"
    TierBenefitDto:
      type: object
      properties:
        grade:
          type: string
        context:
          type: string
...